Use a timeout when getting the Jupyter URL
Wait for a specified number of seconds for the juputer-notebook command
to output the URL instead of assuming it will be in the first few lines.

The timeout can be specified in the config and defaults to 15 seconds.

Also, handle unicode output from the jupyter-notebook command using the
system default encoding.

Fixes #2