Increase default timeout to 30 for getting URL
Wait up to 30 seconds by default for the Jupyter notebook command to
start up and output the URL to use.

Also use subprocess.call instead of check_call when killing the notebook
process and check the output. Previously this could fail and raise an
Exception making the underlying timeout error more difficult to spot.