Add an Excel macro for opening a Jupyter notebook
The new "OpenJupyterNotebook" macro can be called from VBA using
Application.Run and takes a single optional argument.

If passed a path of a notebook file that notebook will be opened. If
passed a directory then Jupyter will be started in that directory.

Also addded a new 'disable_ribbon' option to the config so the macro can
be used without adding anything to the ribbon.

Fixes #6