Fix a problem using Excel from the Windows Store
If Excel has been installed from the Windows Store and the Jupyter
runtime folder is in AppData then use a different path for the
connection files.

This is because Windows Store apps see a different AppData folder than
non-Windows Store apps and so if the connection file is written there
then the Jupyter process can't open it.