Update to work with latest Jupyterlab version
Jupyterlab has changed from using NotebookApp to ServerApp, which
requires a different method of connecting to the IPython kernel.

Also the Jupyterlab entry point has moved, and now displays a console
window if imported directly as it uses subprocess during module import.
Loading the entry point has been moved to the child process to avoid
this.